lambda/function/requirements.txt
aws-xray-sdk==2.9.0
beautifulsoup4==4.10.0
boto3==1.21.36
botocore==1.24.36
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cssselect==1.1.0
cssutils==2.4.0
future==0.18.2
idna==3.3
importlib-metadata==4.11.3
jmespath==1.0.0
jsonpickle==2.1.0
lxml==4.8.0
premailer==3.10.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.2
typing_extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.0
yagmail==0.15.277
zipp==3.8.0
